refactor: reorganize parallel executor tests
